+++ scripts/post-build.js
@@ -17,3 +17,16 @@ utils.processTextFile("everything.d.ts", (text) => {
 
   return text + mark
 })
+
+
+//--------------------------------------------------------------
+// inactive ai1.js when CommonJS or AMD is presented
+
+utils.processTextFile("ai1.js", (text) => {
+  return text.replace(/(\){).+?(\w+\(\w+\.HyperMD=)/, (_, lead, tail) => {
+    return lead +
+      `(("object"==typeof exports&&"undefined"!=typeof module)||("function"==typeof define&&define.amd))?` +
+      `console.error("Don't use HyperMD ai1.js with any Module Loaders!\\nPlease use everything.js, or load modules on demand."):` +
+      tail
+  })
+})